Moved EHS implementation from main to it's own file
